Refactor XORTrialGradClipping into OneVarTrialGradClipping
